+++ app/src/main/java/com/ikarimeister/loginapp/domain/model/Profile.kt
@@ -1,0 +1,3 @@
+package com.ikarimeister.loginapp.domain.model
+
+data class Profile(val username: Email, val token: Token)
