+++ app/src/main/java/com/ikarimeister/loginapp/domain/model/StorageError.kt
@@ -1,5 +1,5 @@
 package com.ikarimeister.loginapp.domain.model
 
 sealed class StorageError
-object TokenNotFound : StorageError()
+object DataNotFound : StorageError()
 data class UnknownStorageError(val t: Throwable) : StorageError()
