+++ app/src/main/java/com/ikarimeister/loginapp/domain/model/Token.kt
@@ -1,0 +1,3 @@
+package com.ikarimeister.loginapp.domain.model
+
+inline class Token(val value: String)
